requirements.txt
# requirements.txt
flwr>=1.0.0
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
pyyaml>=6.0
tqdm>=4.64.0
psutil>=5.9.0
scikit-learn>=1.1.0
pandas>=1.4.0